general: refactoring GameCollections
